requirements.txt
aniso8601==7.0.0
attrs==19.1.0
bokeh==0.12.4
certifi==2018.8.24
chardet==3.0.4
ckanapi==4.3
Click==7.0
dnspython==1.16.0
docopt==0.6.2
eventlet==0.25.0
Flask==1.1.1
flask-restplus==0.13.0
geomet==0.2.0.post2
greenlet==0.4.15
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.17.0
pandas==0.25.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.7.3
python-slugify==3.0.3
pytz==2019.2
PyYAML==5.1.2
rdflib==4.2.2
requests==2.22.0
rfc3987==1.3.8
six==1.12.0
SPARQLWrapper==1.8.4
text-unidecode==1.2
tornado==6.0.3
urllib3==1.25.3
Werkzeug==0.15.5
